+++ src/pyEQL/engines.py
@@ -17,7 +17,6 @@ from phreeqpython import PhreeqPython
 
 import pyEQL.activity_correction as ac
 from pyEQL import ureg
-from pyEQL.salt_ion_match import Salt
 from pyEQL.utils import standardize_formula
 
 # These are the only elements that are allowed to have parenthetical oxidation states
@@ -321,12 +320,9 @@ class NativeEOS(EOS):
         # identify the predominant salt that this ion is a member of
         salt = None
         rform = standardize_formula(solute)
-        for v in solution.get_salt_dict().values():
-            if v == "HOH":
-                continue
-            if rform == v["cation"] or rform == v["anion"]:
-                del v["mol"]
-                salt = Salt.from_dict(v)
+        for d in solution.get_salt_dict(cutoff=0.0).values():
+            if rform == d["salt"].cation or rform == d["salt"].anion:
+                salt = d["salt"]
                 break
 
         # show an error if no salt can be found that contains the solute
@@ -509,12 +505,8 @@ class NativeEOS(EOS):
         # loop through all the salts in the solution, calculate the osmotic
         # coefficint for each, and average them into an effective osmotic
         # coefficient
-        for d in solution.get_salt_dict().values():
-            item = Salt(d["cation"], d["anion"])
-            # ignore HOH in the salt list
-            if item.formula == "HOH":
-                continue
-
+        for d in solution.get_salt_dict(cutoff=0.0).values():
+            item = d["salt"]
             # determine alpha1 and alpha2 based on the type of salt
             # see the May reference for the rules used to determine
             # alpha1 and alpha2 based on charge
@@ -584,8 +576,8 @@ class NativeEOS(EOS):
 
         # use the pitzer approach if parameters are available
         pitzer_calc = False
+        param = None if salt is None else solution.get_property(salt.formula, "model_parameters.molar_volume_pitzer")
 
-        param = solution.get_property(salt.formula, "model_parameters.molar_volume_pitzer")
         if param is not None:
             # determine the average molality of the salt
             # this is necessary for solutions inside e.g. an ion exchange